+++ database/migrations/2024_10_17_102033_create_sessions_table.php
@@ -26,6 +26,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('sessions');
+        if (Schema::hasTable('sessions')) { // Check if the table exists
+            Schema::dropIfExists('sessions');
+        }
     }
 };
