+++ database/migrations/2024_10_30_131511_create_filereturns_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('filereturns');
+        if (Schema::hasTable('filereturns')) { // Check if the table exists
+            Schema::dropIfExists('filereturns');
+        }
     }
 };
