+++ database/migrations/2024_10_31_104503_create__admins_table.php
@@ -26,6 +26,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('_admins'); // Drop the _admins table
+        if (Schema::hasTable('_admins')) { // Check if the table exists
+            Schema::dropIfExists('_admins');
+        }
     }
 };
