+++ database/migrations/2024_10_31_152054_create_contactus_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('contactus');
+        if (Schema::hasTable('contactus')) { // Check if the table exists
+            Schema::dropIfExists('contactus');
+        }
     }
 };
