+++ database/migrations/2025_01_30_121937_create_feedbacks_table.php
@@ -27,8 +27,10 @@ return new class extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
-        Schema::dropIfExists('feedbacks');
+        if (Schema::hasTable('feedbacks')) { // Check if the table exists
+            Schema::dropIfExists('feedbacks');
+        }
     }
 };
